7.0.2 - Sprite CORS support, GPU and Tab adjust
IMPROVEMENT can now load Sprite with CORS (cross domain) when using JSON
parameter
JSON was passing string version of URL to Sprite which voided the image
crossdomain settings
Also, ZIM now pre-parses the images array in the JSON to match
frame.asset() images
So the exact path is no longer needed in the JSON - just a matching
filename will do
This is handy as TexturePacker, etc. often export with just the filename
and previously, this would not run if the SpriteSheet file were stored
in an assets folder for instance
unless the JSON was modified to reflect the path - now, this is not
necessary.
Adjusted gpu setting of Frame to properly make Bitmaps - patched 7 and
7.01
BREAK added a currentSelected parameter after currentEnabled for ZIM Tab
currentEnabled now only determines if tab can be clicked again when
already selected
if true, this will call the change event even though the selectedIndex
has not changed
The currentSelected defaults to true to highlight the selected tab after
clicking
if set to false then the tab does not stay highlighted - good for button
bars - IMPROVEMENT
Version 7 tried to not highlight when currentEnabled was true but this
broke the Pad
so the functionality was split into two parameters
setting currentSelected to false will automatically set currentEnabled
to true
IMPROVEMENT Added bounds transfer to when container is cached -
apparently this is fixed in next CreateJS version