ZIM 7.1.0 - couple big parameter name changes
Updated animate() to use a props parameter rather than an obj parameter.
Changed the existing props parameter to cjsProps as that is for older
transition values.
obj still works for backwards compatibilty but has been depreciated.
This is a conceptual change - props better describes that these are the
properties we are animating.
It will be a fairly wide-spread change so there will be many tutorials
slightly out of date - oh well.
Added a Boundary() class to the code module to store x, y, width and
height
This can be used in place of a createjs.Rectangle in most cases.
The createjs Rectangle is still available and has more features.
This is to avoid confusion between the ZIM Rectangle and the CreateJS
Rectangle.
The ZIM Rectangle is a DisplayObject whereas the CreateJS Rectangle is a
data class.
The Boundary class can be used to specify a boundary for drag(),
gesture() and for a Physics world.
Depreciated the rect parameter for drag() and gesture() and replaced it
with a boundary parameter.
The rect will still work for backwards compatibility
Changed dragRect() and gestureRect() to dragBoundary() and
gestureBoundary()
the old method names are still available but are depreciated.
These methods are used to dynamically change the boundaries - which is
fairly rare.
The change is to match the new Boundary class for specifying a data
rectangle.
Thank you to Bracer Jack for the prompting and confirming our nagging
concern over two Rectangle classes.
Added a Point() class to the code module to store simple x and y
properties.
This can be used in place of a createjs.Point in most cases.
The createjs Point is still available and has more features.
Added a paused property to all DisplayObjects
This will be undefined if the object has not been animated
Once animated, paused will be false until the animation is paused at
which point it will be set to true
If the animation ends or is stopped, the pause property will be set to
null.
This will allow the animation to be toggled with
obj.pauseAnimate(!obj.paused)
IMPROVEMENT - added circularBounds parameter to gesture to keep rotated
circular objects within boundaries
Added a DEFAULTFONT constant to the controls module below ACTIONEVENT
(thanks Bracer Jack for the idea)
This can be set to change the font across all components to whatever is
specified
Fonts specified per component will override this default
DEFAULTFONT can be set to a custom font - frame.asset("myFont.ttf") for
instance
IMPROVEMENT- fixed gesture to not move if move is false and are double
pinching
This was a glitch when we introduced surround bounds
Fixed startBounds parameter for drag() when using DUO.
center() and centerReg() will default to the object's parent if there is
a parent and no container is provided
Thanks Bracer Jack for the suggestion
Changed global variable zon to be true by default to show messages from
ZIM
In CodePen the libraries are passed in separately, so zon is missing
from the code
and people were not receiving ZIM messages like bad parameter warnings,
etc.
So the default now is true and zon must be set to false to suppress
messages.