7.3.0 ProgressBar, Colors, Button Changes, etc.
PATTERNS
Introduced ZIM Pizzazz 03 - custom patterns and animated patterns
These can be passed in to Button, Pane and ProgressBar objects
using the backing and rollBacking, toggleBacking and rollToggleBacking,
waitBacking and rollWaitBacking parameters
at which point the pattern will be masked to the button, bar, etc. size.
The patterns can also be used on their own or masked by Pizzazz 01
shapes or your shapes, etc. using setMask()

COLORS
Added Frame colors to zim namespace - and also global namespace with
default zns set to false
This means to use zim colors just use black, pink, faint, blue, clear,
etc. with no quotes

PROGRESS BAR
Added a ProgressBar() class with default "circle" barType - also
"rectangle" for traditional bar
BREAK Added progress parameter to Frame and LoadAssets after the path
parameter
This can receive a ProgressBar or a Waiter.
BREAK Due to the new Frame progress parameter, it is important to
specify colors as Frame Parameters
so that the colors show when the progress is running - so these have
been moved to before the assets parameter
Loading now defaults to XHR as true if there is a ProgressBar - this
gets load data from preloading
Adjusted Waiter to not require a stage so it can be made and passed in
to progress parameter of Frame and loadAssets()
which may happen before there is a stage in the case of the Frame
progress parameter.

SERIES
IMPROVEMENT Added a makeSeries() function to the Code module that
returns a function which returns values in series
This can be used with ZIM VEE (zik) values to force sequential values
from an array (passed to makeSeries)

TILE CHANGES - IMPROVEMENT / BREAK
Adjusted Tile() so registration and origin are at top left corner of
bounds
Previously these were at the first element's registration point - which
was odd for tiling centerReg objects like circles
Added optional colSize, rowSize, align and valign parameter after
spacings
Moved count parameter to after cols and rows - really tired of adding
null for count all the time
This was added to match the format of importing a SpriteSheet which
often has a count that does not match the tile count
So SpriteSheet will remain as is with the count before spacing.  And
Tile will have the count after spacing.
Added a clone parameter (default true) to end of Tile
Set this to false avoid cloning a series unnecessarily
Added a placeTextColor parameter to Keyboard (BREAK)
Fixed a problem with the icons - required cloning the icons as the board
is remade during state changes

DEFAULTFONT UPDATE - thanks Bracer Jack for the prompting
IMPROVEMENT Applied DEFAULTFONT size and color properties to:
Button, RadioButton, CheckBox, Pane, Tabs, Pad, Stepper, ProgressBar,
Keyboard
Initially, the font (family) and various other properties were applied
but size and color were dictated by the component
Now, all the properties of the default font settings are applied which
could BREAK the look of the app if you were using DEFAULT FONT (only a
few versions old)
These properties include:
size, font, color, rollColor, shadowColor, shadowBlur, fontOptions,
outlineWidth, outlineColor, padding, shiftHorizontal, shiftVertical
align, valign are available for the Pane

BUTTON UPDATES
The Button now has complete parameters and properties for
1. label (text), backing, rollBacking, icon, rollIcon
2. toggle (text), toggleBacking, rollToggleBacking, toggleIcon,
rollToggleIcon
3. wait (text), waitBacking, rollWaitBacking, waitIcon, rollWaitIcon
The backing and icon parameters are read only references to current
objects
Use the new setBacking(type, obj) and setIcon(type, obj) to manually
adjust these at any time
This makes sure that old references are removed and any patterns are
applied properly
Added a rolled property to Button that is true when button is rolled
over
Added a wait() method to force the wait state of the button (similar to
existing toggle() method for toggle state)
Adjusted Button so gradient and gloss work with custom backings

MISC
Added expand() to close button on Pane
Added chainable setBounds() method to Container and all extending
container that overrides the CreateJS setBounds()
Also supports alternate parameters - (size) (width, height) or (x, y,
width, height)