7.3.2 - Window TitleBar, animate Override, etc.
IMPROVEMENT Added an optional titleBar on the top of Window like there
is in Pane to drag and close and have a title...
also added properties to access titleBar and its parts and the close
button - added a "close" event
IMPROVEMENT / BREAK changed the names of Window parameters for indicator
to scrollBar - thanks Bracer Jack
IMPROVEMENT / BREAK inserted scrollBarH and scrollBarV parameters
default true to show scrollBars if scrolling is needed
BREAK called Pane() bar parameters titleBar for consistency with new
Window titleBar parameters
Window has scrollBar parameters too so needed to distinguish a titleBar
rather than use bar.
BREAK Changed the Pane() name of the barTitle parameter to titleBarLabel
BREAK Adjusted ColorPicker to set greyPicker, alphaPicker, buttoBar to
false by default if one row of colors
Added a read only colors property to the ColorPicker that returns the
array of colors
IMPROVEMENT Added "rgba" option to convertColor()
and BREAK changed second parameter to be toColorType with default of
"hex" and other values of "string" and "rgba"
added an alpha parameter for alpha of RGBA conversion
So convertColor(blue, "rgba", .1) would convert the ZIM blue to
rgba(80,196,183,0.1)
IMPROVEMENT / BREAK animate() override now defaults to false instead of
true
except if animating an x or y value and the object is currently
animating in the matching x or y value
In this case, all earlier tweens will be overridden along with the x and
y unless override is set to false
This forces the developer to use two separate tweens if the desired
override settings are different.
Previously, all tweens on an object defaulted to override true - to
solve almost exclusively an x and y animation issue
Adjusted Emitter() to centerReg() but not add as this was overriding the
positioning of the emitter.