ZIM 7.4.0 - Default Style, toggle, backgrounds
DEFAULTSTYLE
IMPROVEMENT Added DEFAULTSTYLE constant similar to the recently added
DEFAULTFONT constant
DEFAULTSTYLE can be used to set any parameter on a components
Examples would be color, rollColor, etc. on the Button components
If the component has a property that can be set by DEFAULTSTYLE then it
checks there first
otherwise it uses the original default for the component.
DEFAULTSTYLE = {color:red, corner:0, step:1, useTicks:true}

IMPROVEMENT for both DEFAULTSTYLE and DEFAULTFONT:
You can now optionally set defaults specifically for types of components
For instance:
DEFAULTFONT = {color:{Stepper:white}}; // will set Stepper fonts white
by default - all others get original defaults
DEFAULTFONT = {color:{Stepper:white, Default:red}}; // will set Stepper
default to white and other component defaults to red
DEFAULTFONT = {color:{Stepper:white, Default:red, Button:"ignore"}}; //
will ignore Button so it gets its original Label defaults
DEFAULTSTYLE =  {color:red, corner:{Button:20, ProgressBar:15}}
Remember - any set object or function that accepts ZIM DUO can receive a
shared custom config objects
These can be merged see merge() with other custom config objects to make
a cascading system like CSS
When doing this, make sure to copy() any config objects with display
objects inside.
Also set the clone parameter of copy() to true

// TOGGLE
There are objects that have control interfaces that can be shown or
hidden:
Grid, Guide, Layout, Blob, Squiggle and objects with outline() or
transform()
We have added a toggle(state) method to these
This has also been added to the Manager class which gives it to the
GridManager, GuideManager, LayoutManager and TransformManager
Using toggle() will show the control (if hidden) or hide the control (if
showing)
You can also use toggle(true) to specifically show and toggle(false) to
specifically hide
A read-only toggled property has been added that is true if visible and
false if not
BREAK - the transform(), Blob() and Squiggle() toggle parameter and
property have been changed to allowToggle
Added the toggle system to Keyboard and ColorPicker, ProgressBar and
Waiter (Pane already had one)
Many of these have hot keys to toggle or press on / press off to toggle
and these still work
They also have show() and hide() and visible properties or
showControls() and hideControls() and controls property
These will still work as well.
Added a show(), hide() to ColorPicker to allow it to act like a pop-up
like Pane, Waiter, ProgressBar and Keyboard
Added an optional container parameter to ColorPicker - ColorPicker can
still be added with addTo(), center(), etc.
but if it is being used as a pop-up, the show() and hide() or toggle()
should be used.

// BACKINGS and BACKGROUNDS
The term backing is now consistently used for a custom DisplayObject
added to a component
(such as a Pizzazz shape or pattern)
The term background is now consistently used for a built in background
rectangle
These are specified with the backgroundColor parameter for Label,
Indicator, etc. objects
In the past, both backingColor and backgroundColor were used which was
inconsistent.
BREAK Changed Indicator, Tabs, ColorPicker and TextArea backingColor to
backgroundColor
Renamed the backing properties of these to background to access the
rectangle

// GENERAL
BREAK - changed the name of the Label backgroundCorner to just corner
(easier to style)
BREAK - Added a corner property to Indicator after backgroundColor
BREAK - Added a borderColor and borderWidth to Keyboard - so changes
parameter order
BREAK - added a borderWidth property to Stepper and Indicator - this is
after borderColor so breaks parameter order
Also made border work on Indicator background
Adjusted copy() to not break when cloning a null value
IMPROVEMENT Fixed Button rollPersist to work if release outside of
Button
Added indicatorColor property to CheckBox and RadioButtons that defaults
to color
Adjusted ColorPicker so color text does not shift when picked (due to
changes in centerReg() functionality)
IMPROVEMENT - clone() now clones the type property
so if you change the type of a Rectangle and clone it, the clone will
have the changed type
Handy for passing an object with a custom type="Pattern" in for backings
in DEFAULTSTYLE
Adjusted Keyboard so selectedIndex starts at the end of the label text -
also fixed backspace at 0 index with existing text bug
IMPROVEMENT - updated Button so when it has a backing or icon,
it removes the rollover state when enabled is set to false