ZIM OCT (8) Styles for the Components
ZIM OCT (8) ~ added 12k
STYLE
Replaced DEFAULTFONT with a general STYLE system. IMPROVEMENT
STYLE can be used to set any parameter on a component.
These are applied at the time the components are made.
They are cascading with each level overriding the previous level:

1. GENERAL: any style can be specified in general
corner:30 will make all corners default to 30
2. TYPE: styles for component type can be set to override the general
styles
Button:{corner:0} will make all button corners default to 0
3. GROUP: styles for a group can be set to override the type styles
homePage:{corner:20} will make all components of that group default to
20
4. OBJECT: styles applied as parameters to the component object override
all other styles
new Button({corner:40}) will make this button have a corner of 40

STYLE = {
corner:0,
type:{
Button:{corner:20, outline:true},
Dial:{backgroundColor:red, scale:2}
},
group:{
homePage:{corner:40}
}
}

See: https://zimjs.com/style.html for an example

TRANSFORM STYLES
Transformations can also be applied (all are numbers / visible is a
Boolean):
x, y, rotation, alpha, scale, scaleX, scaleY, regX, regY, skewX, skewY,
visible.

FUNCTION STYLES
Popular functions have been added - parameters may be provided in future
versions
but for now, use a value of true for default functionality.
drag, transform, gesture, and outline - all will operate on the
component container (currentTarget).

EXCLUSION
A value of ignore can be used to exclude any earlier styles and return
to the original default.
ignore is a ZIM global variable - if zns is true then use zim.ignore or
just "ignore".
An ignoreStyles parameter has been added to all components. Set to true
to ignore all styles.

GROUPS
Added a group parameter to the end of the parameters of all components.
This parameter accepts a string or a comma delimited string of multiple
groups.
The term "class" was not used as all objects in code are made from
classes.
The group of components can then be targeted in the group section of
STYLE.
A group can contain just one component and act like an ID in CSS.

COLORS
Colors have been reworked for the components (they stay the same for
shapes)
Basically, color is now used for text and backgroundColor for the
backgrounds - like CSS.

color - text (and rollColor)
backgroundColor - background of component (and rollBackgroundColor)
foregroundColor - foreground of component
borderColor - border on background and foreground
indicatorColor - indicator of component

SUMMARY OF PARAMETER CHANGES - BREAK
Most non-Label color and rollColor parameters and properties
have been changed to backgroundColor and rollBackgroundColor
color and rollColor have been kept when the component has text
and these now refer to the text - specifying a Label will override
these.
Added backgroundColor to CheckBox and RadioButtons for the box and
circle
The Tabs and Pad offColor is now the text color
and they get an offBackgroundColor to specify non-selected buttons

Exceptions:
Most components have their color parameter changed to backgroundColor.
The exceptions are those which shift to a foregroundColor parameter:
Waiter, ProgressBar, Indicator has its color parameter changed to
foregroundColor.
The color parameter remains on the ProgressBar for its text color.
Waiter has its foregroundColor moved to before its backgroundColor - for
consistency

Keyboard:
Changed Keyboard parameters so text gets color and backgrounds get
backgroundColor
There are a bunch of these so here they are - old on the left and new on
the right:
color to backgroundColor
textBackgroundColor to color
shiftHoldColor to shiftHoldBackgroundColor
shiftBackgroundColor to shiftColor
placeColor to placeBackgroundColor
shiftColor to shiftBackgroundColor
placeTextBackgroundColor to placeColor

// TOGGLE
There are objects that have control interfaces that can be shown or
hidden:
Grid, Guide, Layout, Blob, Squiggle and objects with outline() or
transform()
We have added a toggle(state) method to these
This has also been added to the Manager class which gives it to the
GridManager, GuideManager, LayoutManager and TransformManager
Using toggle() will show the control (if hidden) or hide the control (if
showing)
You can also use toggle(true) to specifically show and toggle(false) to
specifically hide
A read-only toggled property has been added that is true if visible and
false if not
BREAK - the transform(), Blob() and Squiggle() toggle parameter and
property have been changed to allowToggle
Added the toggle system to Keyboard and ColorPicker, ProgressBar and
Waiter (Pane already had one)
Many of these have hot keys to toggle or press on / press off to toggle
and these still work
They also have show() and hide() and visible properties or
showControls() and hideControls() and controls property
These will still work as well.
Added a show(), hide() to ColorPicker to allow it to act like a pop-up
like Pane, Waiter, ProgressBar and Keyboard
Added an optional container parameter to ColorPicker - ColorPicker can
still be added with addTo(), center(), etc.
but if it is being used as a pop-up, the show() and hide() or toggle()
should be used.

// BACKINGS and BACKGROUNDS
The term backing is now consistently used for a custom DisplayObject
added to a component
(such as a Pizzazz shape or pattern)
The term background is now consistently used for a built in background
rectangle
These are specified with the backgroundColor parameter for Label,
Indicator, etc. objects
In the past, both backingColor and backgroundColor were used which was
inconsistent.
BREAK Changed Indicator, Tabs, ColorPicker and TextArea backingColor to
backgroundColor
Renamed the backing properties of these to background to access the
rectangle

// GENERAL
BREAK - changed drag parameters to draggable - as conflicts with drag
function style
BREAK - changed the name of the Label backgroundCorner to just corner
(easier to style)
BREAK - Added a corner property to Indicator after backgroundColor
BREAK - Added a borderColor and borderWidth to Keyboard - so changes
parameter order
BREAK - added a borderWidth property to Stepper and Indicator - this is
after borderColor so breaks parameter order
Also made border work on Indicator background
Adjusted copy() to not break when cloning a null value
IMPROVEMENT Fixed Button rollPersist to work if release outside of
Button
Added indicatorColor property to CheckBox and RadioButtons that defaults
to color
Adjusted ColorPicker so color text does not shift when picked (due to
changes in centerReg() functionality)
IMPROVEMENT - clone() now clones the type property
so if you change the type of a Rectangle and clone it, the clone will
have the changed type
Handy for passing an object with a custom type="Pattern" in for backings
in STYLE
Adjusted Keyboard so selectedIndex starts at the end of the label text -
also fixed backspace at 0 index with existing text bug
Fixed animate() when running a series with as a global function with no
target
the last change for overriding by default only if x and y are animating
- was expecting a target - fixed now...
IMPROVEMENT - updated Button so when it has a backing or icon,
it removes the rollover state when enabled is set to false
Adjusted backing on Label to mask backings of type="Pattern"

WELCOME TO ZIM OCT!
A kids section is also being added to ZIM Skool for workshops and
secondary school lessons
Please let others know about ZIM and we'd love to hear from you at
https://zimjs.com/slack - very productive!