ZIM 8.3.0 added loc(), path animation, blob...
Added loc() chainable method - pos() no longer positions the
registration point unless the reg parameter is set to true
So loc() will position the registration point at loc(x, y) and default
to add to stage, or choose some other container or add at index, etc.
BUT alternatively, loc can also accept any object with an x and y
property as its first parameter
This could be another Display Object or a zim Point(100, 100) or an
object literal {x:100, y:100}, etc.
When the first parameter is a Display Object, obj.loc() will put its obj
at the same place (registration wise) as the Display Object.
This will default to set localToLocal true for exact match - there is a
localToLocal parameter if this is not desired.
So... loc acts just like the old pos() with the option to
obj.loc(displyObject) to easily match positions
Currently there is no loc style - it is easier to use x and y styles
as in x:10, y:10 rather than loc:{x:10, y:10}

PATH
https://zimjs.com/explore/squiggleAnimate.html
Added animation along the path with ZIM animate - the following props
are added:
path - pass in a Blob or Squiggle to animate along path
this is dynamic so if the Blob or Squiggle changes, the animation will
change
if Blob or Squiggle has lockControls = false, then it is not dynamic
if a set path is desired, then lockControls to save processing
https://zimjs.com/explore/blobAnimate.html
orient - an object to rotate the target towards
if path is being used, orient defaults to orient along the path
can set to false for no rotation
can set to any object with an x, y point such as a circle on the stage -
or {x:0, y:0} or a new zim Point, etc.
the target will then rotate to face the specified location
https://zimjs.com/explore/blobAnimate2.html

BLOB and SQUIGGLE
IMPROVEMENT Added click blob to add control point and ctrl click point
to remove a control point
https://zimjs.com/explore/blobAdd.html
Added an editPoints parameter that defaults to true for the above - or
set to false to not add and remove control points
gave Blob and Squiggle controls a num property that matches their points
array position
Fixed a glitch when setting points property that was introduced in 7.4.0
where we started using the Container's dispose
IMPROVEMENT Added proper color to handle rectangles when using Blob
changeControl
Added segmentPoints and segmentPercents properties to Blob and Squiggle
Added getPointAngle(), getSegmentPoint(), getCurvePoint() methods to
Blob and Squiggle
Added pointAlongCurve, distanceAlongCurve, closestPointAlongCurve
functions to Code module
IMPROVEMENT Blob and Squiggle now persists even with a changed number of
points
Added update event to blob and squiggle for when number of points change
All events on Blob will now be kept but any custom events on the shape
or controls will need to be recreated
as the shape and controls are now different objects. So capture the
update event and remake your events.
Any events on the shape or the controls will be removed before the
objects are replaced with new ones.