ZIM 9.1.2 - Path improvements, Label, Series...
SQUIGGLE AND BLOB UPDATES
Added an interactive parameter to Squiggle and Blob to quickly turn off
controls, drags, selects, etc.
This leaves just the shape and should be set to false when not needing
to interact with shape
Added pointControls and pointCircles properties to Squiggle and Blob to
access these arrays directly
Before we had to go through pointObjects[0][0] to get to control
container and pointObjects[0][1] to get to circles
Animation should be done on the pointControls but immediate placement
should be done on the pointCircles
This is because when dragged, the circle is moved but the control is
left in place until mouseup
Added a transformPoints() function in the Code module to scale, rotate
or position points in a Squiggle or Blob format
Added a transformPoints() method to Squiggle and Blob that use the
transformPoints function to change their points
This allows you to scale or rotate Squiggle or Blob points without
affecting their control points or stroke thickness
EXAMPLE
// Transform the original points of a Blob
// If you rotate or scale, this affects the control points - the little
rectangles rotate or they scale
// To avoid this, the points themselves can be transformed (scaleX,
scaleY, scale, rotation, x, y)
// This makes a square and scales it bigger without affecting control
size or stroke size (if there were a stroke)
// Note the default number of points is 4 but they are arranged at the
top, bottom and sides - so would make a diamond with just
controlType:"none"
new Blob({controlType:"none"}).transformPoints("rotation",
45).transformPoints("scale", 2).center();
END EXAMPLE
Fixed a glitch when unpausing a dragged object along a path
Made objects follow a Squiggle or Blob path as the path is being moved
or manipulated IMPROVEMENT (although also patched in ZIM 9.1.0)
This results in a better experience when editing live animations

EMITTER
Adjusted Emitter to work if not added to stage or parent.
Changed Emitter pause() and paused to emitterPause() and emitterPaused -
to not conflict with animate() pause() and paused
as sometimes we animate the Emitter.  BREAK

SERIES
Made series() hold just regular parameters or a single parameter of an
array.
series(red, green, blue) is now the same as series([red, green, blue])

LABEL IMPROVEMENT
Added a "baseline" value for valign parameter of Label that puts the
registration of text at the baseline
Added labelWidth and labelHeight parameters to Label - use them both to
override font size and fit text to dimensions provided.
labelHeight on its own does nothing.  labelWidth on its own is the same
as lineWidth - it will wrap a the provided value.