ZIM 9.2.0 - Pen, List, Toggle, LabelOnPath, lots!
9.2.0
PEN
https://zimjs.com/pen.html
Handles dynamic drawing with a set of different penTypes
You can drag() or gesture() pen, move it with a MotionController(),
animate() it, animate() it along a Squiggle or Blob path, etc.
Holding down the CTRL key will allow the pen to move without drawing

LABELONPATH
https://zimjs.com/explore/labelonpath.html - thanks KV for the request!
Makes a label along a path of a Squiggle or Blob - which can be
interactive, fixed, toggled or hidden
A list of percentages for where the letters are can be provided to move
around letters

TOGGLE
https://zimjs.com/explore/toggle.html - thanks Andi Erni for the idea
and code start.
Added a Toggle() component after the RadioButtons
Parameters: width, height, label, startToggled, backgroundColor, margin,
indicatorType, indicatorColor, toggleBackgroundColor,
color, borderColor, borderWidth, corner, indicatorCorner, shadowColor,
shadowBlur, time, labelLeft, style, group
Gives a change event and check the toggled (or selected) property -
works with STYLE and Accesibility

BUTTON AND TABS - IMPROVEMENT
https://zimjs.com/explore/verticalTabs.html
Added align, valign, indent, indentHorizontal, indentVertical parameters
to end of Button and Tabs
BREAK - added a vertical parameter to Tabs just before spacing set to
true to make vertical tabs
BREAK - flatBottom parameter of Tabs is changed to base with default
"none" when no corner or corner is an array
and more values of "bottom" (default when corner and not vertical),
"left" (default when corner and vertical), "top", "right"
corner will then be the opposite corners to the base
Fixed currentSelected to not show selected value - got lost in color /
backgroundColor change over
BREAK - removed flatBottom from Rectangle as corner is now configurable
with Array and Tabs use base rather than flatBottom
Note - this breaks any traditional parameter order pointing to dashed.
Added labelIndent, labelIndentHorizontal, labelIndentVertical parameters
to Tabs

**** Added change() and noChange() chainable method to capture change
events on components

FLATBOTTOM - REMOVED - BREAK
The flatBottom parameter has been removed as corner now accepts an array
and Tabs use base for more flexible horizontal vertical arrangement
So the parameter order of Rectangle, Button and Loader from dashed on
has been changed

LIST
https://zimjs.com/explore/list.html
https://zimjs.com/explore/listObjects.html
Added a ZIM List component for a vertical or horizontal list of items -
thanks Andi Erni for the prompting
This is really a zim.Tabs object inside a zim.Window object.
The list can be strings, numbers or Label objects and these are added to
Tabs buttons.
The list can also include any DisplayObject with bounds (which most ZIM
objects have except a Shape needs bounds set manually with setBounds()).
If the object is not a string, number or Label then selection will not
highlight and currently animateTo() may not work if size is different.
Items can be added or removed from the list with addAt() and removeAt()
You can animateTo() and index as well - not working properly yet for
when custom objects are in the list

IMPROVEMENT MOUSE POSITION ON TOUCHSCREEN
USE Mouse Event Object: e.stageX AND e.stageY - NOT stage.mouseX and
stage.mouseY - these last two do not work with touch screen.
We have 100 places where we get mouse positon and almost all are working
fine.
But we missed one for the initial point down for dragging with slide -
this has been fixed
Went through docs and fixed all examples using stage.mouseX, etc.
Fixed minor places in Stepper and Parallax as well.

Added default particles to Emitter.
IMPROVEMENT - Squiggle and Blob when animating rotation of points now
adjusts animation along path correctly - for animated paths (interactive
was fine)
BREAK - Indicator parameters have changed to foregroundColor,
backgroundColor (for not selected) and backdropColor for the backdrop
(which is often left out)
The background property has been changed to backdrop.  Pane and Tabs
also have a backdrop - it is the background behind the background ;-)
So the layers are selectedColor (font), color (font), foregroundColor,
selectedBackgroundColor, backgroundColor, backdropColor - sigh.
Added onTop and interactive properties to Squiggle and Blob
Added a getAngle parameter to Squiggle and Blob getCurvePoint() and the
CODE module pointAlongCurve() function
These now return an angle at point property along with x, y, etc.
Fixed a bug in pos() when bounds set to negative x and y
Made align:"middle" work as well as align:"center" on label
Adjusted ColorPicker to accept borderWidth and borderColor styles - for
color rectangles or circles
Added borderColor and borderWidth styles for Tabs
Fixed glitch in Swiper to work properly with stage as the swipeOn
parameter.
Fixed glitch in MotionController introduced in 9.1.0 with accelerator
(put brackets around ternary when used in calculation)
BREAK - changed mouseDownIncludes parameter and property of
MotionController to mousedownIncludes.
Added target automatically to mousedownIncludes in MotionController -
this can be removed from the property array if not desired
Fixed flip vertical problem with MotionController - now points the right
way!  We suspect hackers.
Added a clear() method to ZIM Dictionary()
BREAK - adjusted Label to not automatically center on a backing (by
default there is no backing)
backing is added when you add a custom backing DisplayObject
The Label now will align and valign properly on a backing with padding -
as opposed to automatically centering.
BREAK - changed CheckBox check property to indicator and color to
indicatorColor
BREAK - changed RadioButton dots property to indicators
Adjusted margin on CheckBox to be closer to CheckBox when set to 0
(inline with RadioButtons and new Toggle)
Adjusted the right hand and bottom margin of RadioButtons a little.
BREAK for objects with outline(), changed toggle() and toggled to
outlineToggle() and outlineToggled so as to not conflict with other
toggle() and toggled methods and properties
Fixed talk() for Accessibility - talkTag.setAttribute("aria-hidden",
!_aria); seemed to be making it not work anymore?
Added Toggle to Accessibility
Added wait time to sequenceWait call - was taken out during wait updates
Fixed a bug in running a Sprite backwards and not ending at 0 - thanks
Alex Gopher for the find.