9.2.1
Adjusted STYLE, POSREG, DRAGALL and KEYFOCUS to work with zns=true -
thanks Ami for the find.
Changed about 20 references to colors, and zim functions like copy,
merge, etc. where namespaces were not used - this can mess up when zns
is true.
This update needed to be made but could not safely be a patch due to the
number of changes.
Adjusted frame.asset("somesound.mp3").play({loop:-1}) to also be
{loop:true} and added a loopCount property for optional number of loops.
but can also still use loop:-1 or loop:10.
This is consistent with how loop works in animate().  The loop:-1 still
works as well.
Added a password Boolean parameter to TextArea to make the field a
password field (one line input field) - thanks Andi for the suggestion
If animate() is missing props it gives a console message (with zon=true)
and returns the object.