9.3.0 - Pen, Tag, ord, List updates
PEN
https://zimjs.com/explore/pen_undo.html
Re-organized internal pen to accommodate multiple undo and redo and
individual drags
The paper property is now a Container that holds either Bitmaps (if
cache is true) or a shapes (if cache is not true)
Added a draggable parameter (default true) to let pen segments be
draggable
Added an onTop parameter to dictate whether the dragged segment comes on
top
Holding the CTRL key will drag all the segments - actually drags the
paper
Added an deletable parameter (default true) to let user remove segments
SHIFT press on a segment to remove it
CTRL SHIFT press to delete all
paper refers to the Container and is read write to change containers if
desired (draw different layers, etc.)
The shape parameter has been changed to paper - to receive an optional
Container
If no container is provided Container is made and placed one layer under
the pen in the Container that holds the pen
There are no more shape and bitmap properties as there is now a bitmap
or shape per each line drawn - so use paper.getChildAt() if necessary
Added a "stop" event to pen that triggers when the motion has stopped
(after damping)
Pen now has an undo parameter that defaults to 0 - set to a number to
add undo and redo functionality
Added a saveState(obj) method to record a transform change to the paper
or a line segment for an undo state (only needed on manual change)
This by default happens when the stop event triggers
Added an undo() method that goes back one recorded state - can be called
multiple times - but does nothing if past the undo level set in
parameter
Added a redo() method to redo any undo() calls - unless draws.
Added an undoLevels property to get or set the number of undo levels
Added a draggable property to read only if segments are draggable
Added a cropScale parameter that defaults to 1 (stage size) - increase
if you want to drag or move shape and paper
Made pen use localToLocal with shape so can move paper and still sync up
drawing.
Added a delete(index) method to delete a line segment at a given index
(actually sets its alpha to 0 to maintain layers on undo)
Added a deleteSegment(segmentObject) method to delete a line segment
object (actually sets its alpha to 0 to maintain layers on undo)

TAG
https://zimjs.com/explore/tag.html
Added a Tag() class to create and overlay a div tag with a certain
dimension
Any HTML can be added with the add()
Also has tag, innerHTML and style properties
There is a resize() method to be called if manually moved or scaled.
Resizing a fit or outside mode will auto resize.
Pane and Pages will handle adding and removing.
The overlay effect is the same as with TextArea and Loader
So canvas objects will always appear beneath the Tag
and there is a minor lag when resizing windows or dragging a pane, etc.

ORD
Added an ord(num) chainable method to move (relative) layer index
So circle.bot().ord(1) moves up one from bottom and circle.top().ord(-2)
moves two down from top

INHERIT
Added inherit parameter to all classes with style.
This can pass styles on to objects made by the class - used internally
But could be used to pass styles directly into a class as it is created
Made List pass on styles to Window and Tabs through this system
Now setting size:10 on list will change the size of the Label in the
Buttons of the Tabs of the List (unless custom objects in list)

LIST
Fixed List to work properly with padding (and paddingVertical,
paddingHorizontal) - these now default to the spacing in their
direction.
They can be set to 0 to have no space on the ends of the list
Added horizontal slideSnap to Window - was left off
Changed List to default to currentSelected:true
Added first() and last() chainable methods to Tabs and List
Added selected, text and label properties to List
Fixed a slight positioning error in addAt() - no need to re-adjust the
tab size as it is done in tabs
Fixed Tab select and unselect colors now that items can be added to and
removed from tabs - internally, the tabs should only be used at start
Fixed up scroll max to take into account padding

GENERAL
Adjusted getQueryString() to remove + from urlEncoded data
Adjusted Bitmap to clone actual width and height not ignored 100x100
default as that messed up clones.
Adjusted Stepper so setting min and max property limits number and
adjusts greyed arrow if need-be.
Patched 9.2.1 with stepper setting as well.
Made MotionController ignore CreateJS objects which do not have a
hitTestPoint, etc.